Move UseApr enum inside the Tomcat class
See gh-44033
